requirements.txt
numpy == 1.17.2
opencv-python == 4.1.1.26
tensorflow-gpu == 1.9.0
keras == 2.2.5